refactor : 코드리뷰 내용 적용 1. 레이싱 결과를 매번 출력하지 않고, raceResultByStep에 담아두었다가 한번에 출력 2. ResultView를 domain으로 이동, RaceService가 ResultView에 직접 의존하도록 수정
