+++ src/main/kotlin/racingcar/client/RaceClient.kt
@@ -4,9 +4,9 @@ import racingcar.domain.car.Car
 import racingcar.domain.car.Cars
 import racingcar.domain.race.Race
 import racingcar.domain.randomnumber.RandomNumbers
+import racingcar.domain.result.ResultView
 import racingcar.domain.service.RaceService
 import racingcar.view.InputView
-import racingcar.view.ResultView
 
 class RaceClient {
     fun startRace() {
@@ -44,9 +44,9 @@ class RaceClient {
     private fun executeRace(
         tryTime: Int,
         carCount: Int,
-        execute: (Int, (Cars) -> Unit, List<RandomNumbers>) -> Race,
+        execute: (Int, List<RandomNumbers>) -> Race,
     ): Race {
         ResultView.printStart()
-        return execute(tryTime, ResultView::printResult, List(tryTime) { RandomNumbers(carCount) })
+        return execute(tryTime, List(tryTime) { RandomNumbers(carCount) })
     }
 }
